gateway/requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
redis==5.0.1
httpx==0.26.0
prometheus-client==0.19.0
pyjwt==2.8.0
bcrypt==4.0.1
python-dotenv==1.0.0
